requirements.txt
pyspark==3.5.3
pandas
torch
transformers==4.44.2
faiss-gpu
typing_extensions
sentence-transformers

